Add files via upload
ID: CreditCardFraud

Data: http://www.dei.unipd.it/~silvello/FIS2020/CC_frauds.zip

Description: The dataset contains transactions made by credit cards in September 2013 by european cardholders. This dataset presents transactions that occurred in two days, where we have 492 frauds out of 284,807 transactions. The dataset is highly unbalanced, the positive class (frauds) account for 0.172% of all transactions.

The task is to design and build a system that automatically recognizes a fraudulent transaction.


A Python notebook reporting all the steps conducted to implement the system: from the cleaning of data, to the design of the features, the training and validation of the system, and the analysis of the results.